Rename to create
- Rename the new function to `create`
- Make the default export use it
- Add documentation
